[mkcal] Restrict event reminder automatic snoozing to 2 times
Set the timed enabler for maximum number of snoozes that the
event reminder can be automatically snoozed as a result of
user inactivity. If the user does not respond to the event
reminder for a third time, then timed will dismiss the reminder.
The event reminder will be shown a total of three times in case
the user does not react to it.